boolforge refactored
