+++ src/features/boolforge/components/TruthTable.jsx
