+++ src/features/boolforge/components/index.jsx
@@ -1,2 +1,4 @@
 export { Sidebar } from './Sidebar';
 export { RenameModal } from './RenameModal';
+export { CircuitCanvas } from './CircuitCanvas';
+export { CircuitControls } from './CircuitControls';
