+++ src/features/boolforge/hooks/index.jsx
@@ -1,1 +1,5 @@
 export * from './useKeyboardShortcuts';
+export * from './useCircuitState';
+export * from './useCanvasInteractions';
+export * from './useSimulation';
+export * from './useAI';
