+++ src/features/boolforge/utils/constants.jsx
