+++ src/features/boolforge/utils/index.jsx
@@ -1,3 +1,5 @@
 export * from './constants';
 export * from './canvasUtils';
 export * from './gateLogic';
+export * from './circuitHelpers';
+export * from './truthTableHelpers';
